Removed Route class

+++ src/controllers/transactions/transactions.delete.ts
@@ -1,32 +1,34 @@
 import { transactionsRouter } from "..";
 import { getProtectedTransaction } from "../../utils/getProtectedTransaction";
 import { prisma } from "../../server";
-import { Route } from "../../utils/Route";
+import { UnauthenticatedFailure } from "../../utils/Failures";
 
-new Route(transactionsRouter, "/:id").protected.delete(
-  async (user, req, res) => {
-    /**
-     * Get ID from request params
-     */
-    const id = req.params.id;
-
-    /**
-     * Find user's transaction with given ID
-     */
-    const transaction = await getProtectedTransaction(user, id);
-
-    if (transaction.isFailure()) {
-      return transaction;
-    }
-
-    /**
-     * Delete transaction
-     */
-    await prisma.transaction.delete({ where: { id: transaction.value.id } });
-
-    /**
-     * Respond with 204
-     */
-    res.status(204).end();
+transactionsRouter.delete("/:id", async (req, res, next) => {
+  if (!req.data.user) {
+    return next(new UnauthenticatedFailure());
   }
-);
+
+  /**
+   * Get ID from request params
+   */
+  const id = req.params.id;
+
+  /**
+   * Find user's transaction with given ID
+   */
+  const transaction = await getProtectedTransaction(req.data.user, id);
+
+  if (transaction.isFailure()) {
+    return next(transaction);
+  }
+
+  /**
+   * Delete transaction
+   */
+  await prisma.transaction.delete({ where: { id: transaction.value.id } });
+
+  /**
+   * Respond with 204
+   */
+  return res.status(204).end();
+});
